reorganize handler based on oss/proprietary
