+++ berkeley-function-call-leaderboard/bfcl/model_handler/handler_map.py
@@ -1,25 +1,25 @@
-from bfcl.model_handler.arctic_handler import ArcticHandler
-from bfcl.model_handler.claude_handler import ClaudeHandler
-from bfcl.model_handler.cohere_handler import CohereHandler
-from bfcl.model_handler.databricks_handler import DatabricksHandler
-from bfcl.model_handler.deepseek_handler import DeepseekHandler
-from bfcl.model_handler.firework_ai_handler import FireworkAIHandler
-from bfcl.model_handler.functionary_handler import FunctionaryHandler
-from bfcl.model_handler.gemini_handler import GeminiHandler
-from bfcl.model_handler.gemma_handler import GemmaHandler
-from bfcl.model_handler.glaive_handler import GlaiveHandler
-from bfcl.model_handler.gorilla_handler import GorillaHandler
-from bfcl.model_handler.gpt_handler import OpenAIHandler
-from bfcl.model_handler.hermes_handler import HermesHandler
-from bfcl.model_handler.llama_handler import LlamaHandler
-from bfcl.model_handler.mistral_handler import MistralHandler
-from bfcl.model_handler.nexus_handler import NexusHandler
-from bfcl.model_handler.oss_handler import OSSHandler
-from bfcl.model_handler.granite_handler import GraniteHandler
-from bfcl.model_handler.nvidia_handler import NvidiaHandler
-from bfcl.model_handler.glm_handler import GLMHandler
-from bfcl.model_handler.yi_handler import YiHandler
-from bfcl.model_handler.xlam_handler import xLAMHandler
+from bfcl.model_handler.proprietary_model.snowflake import ArcticHandler
+from bfcl.model_handler.proprietary_model.claude import ClaudeHandler
+from bfcl.model_handler.proprietary_model.cohere import CohereHandler
+from bfcl.model_handler.proprietary_model.databricks import DatabricksHandler
+from bfcl.model_handler.oss_model.deepseek import DeepseekHandler
+from bfcl.model_handler.proprietary_model.fireworks import FireworksHandler
+from bfcl.model_handler.proprietary_model.functionary import FunctionaryHandler
+from bfcl.model_handler.proprietary_model.gemini import GeminiHandler
+from bfcl.model_handler.oss_model.gemma import GemmaHandler
+from bfcl.model_handler.oss_model.glaive import GlaiveHandler
+from bfcl.model_handler.proprietary_model.gorilla import GorillaHandler
+from bfcl.model_handler.proprietary_model.openai import OpenAIHandler
+from bfcl.model_handler.oss_model.hermes import HermesHandler
+from bfcl.model_handler.oss_model.llama import LlamaHandler
+from bfcl.model_handler.proprietary_model.mistral import MistralHandler
+from bfcl.model_handler.proprietary_model.nexus import NexusHandler
+from bfcl.model_handler.oss_model.base_oss_handler import OSSHandler
+from bfcl.model_handler.oss_model.granite import GraniteHandler
+from bfcl.model_handler.proprietary_model.nvidia import NvidiaHandler
+from bfcl.model_handler.oss_model.glm import GLMHandler
+from bfcl.model_handler.proprietary_model.yi import YiHandler
+from bfcl.model_handler.oss_model.salesforce import SalesforceHandler
 
 handler_map = {
     "gorilla-openfunctions-v0": GorillaHandler,
@@ -65,8 +65,8 @@ handler_map = {
     "mistral-small-2402-FC-Any": MistralHandler,
     "mistral-small-2402-FC-Auto": MistralHandler,
     "mistral-tiny-2312": MistralHandler,
-    "firefunction-v1-FC": FireworkAIHandler,
-    "firefunction-v2-FC": FireworkAIHandler,
+    "firefunction-v1-FC": FireworksHandler,
+    "firefunction-v2-FC": FireworksHandler,
     "Nexusflow-Raven-v2": NexusHandler,
     "gemini-1.0-pro": GeminiHandler,
     "gemini-1.5-pro-preview-0409": GeminiHandler,
@@ -95,6 +95,6 @@ handler_map = {
     "nvidia/nemotron-4-340b-instruct": NvidiaHandler,
     "THUDM/glm-4-9b-chat": GLMHandler,
     "yi-large-fc": YiHandler,
-    "Salesforce/xLAM-1b-fc-r": xLAMHandler,
-    "Salesforce/xLAM-7b-fc-r": xLAMHandler
+    "Salesforce/xLAM-1b-fc-r": SalesforceHandler,
+    "Salesforce/xLAM-7b-fc-r": SalesforceHandler,
 }
