+++ berkeley-function-call-leaderboard/bfcl/model_handler/oss_model/gemma.py
@@ -1,4 +1,4 @@
-from bfcl.model_handler.oss_handler import OSSHandler
+from bfcl.model_handler.oss_model.base_oss_handler import OSSHandler
 from bfcl.model_handler.utils import ast_parse
 import re
