+++ berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/claude.py
@@ -5,7 +5,7 @@ import time
 from anthropic import Anthropic
 from anthropic.types import TextBlock, ToolUseBlock
 from bfcl.model_handler.constant import GORILLA_TO_OPENAPI, DEFAULT_SYSTEM_PROMPT
-from bfcl.model_handler.handler import BaseHandler
+from bfcl.model_handler.base_handler import BaseHandler
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.model_handler.utils import (
     ast_parse,
