+++ berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/cohere.py
@@ -1,6 +1,6 @@
 import os
 
-from bfcl.model_handler.handler import BaseHandler
+from bfcl.model_handler.base_handler import BaseHandler
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.model_handler.utils import (
     func_doc_language_specific_pre_processing,
