+++ berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/mistral.py
@@ -1,4 +1,4 @@
-from bfcl.model_handler.handler import BaseHandler
+from bfcl.model_handler.base_handler import BaseHandler
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.model_handler.constant import (
     DEFAULT_SYSTEM_PROMPT,
