+++ berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/nvidia.py
@@ -1,6 +1,6 @@
 import time, os
 from openai import OpenAI
-from bfcl.model_handler.handler import BaseHandler
+from bfcl.model_handler.base_handler import BaseHandler
 from bfcl.model_handler.model_style import ModelStyle
 from bfcl.model_handler.utils import ast_parse
 from bfcl.model_handler.utils import (
